third_party/blink/web_tests/external/wpt/css/css-contain/content-visibility/content-visibility-svg-path-expected.txt

This is a testharness.js-based test.
[FAIL] checkEnclosure() should return true in a c-v:hidden subtree.
  assert_true: enclosure expected true got false
[FAIL] getEnclosureList() should return items in a c-v:hidden subtree.
  assert_greater_than: intersection expected a number greater than 0 but got 0
Harness: the test ran to completion.
